refactor: split release parser helpers
